src/requirements.txt
anyio==3.6.2
boto3==1.26.91
botocore==1.29.91
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
fastapi==0.94.0
google-api-core==2.11.0
google-auth==2.16.2
google-cloud-pubsub==2.15.1
googleapis-common-protos==1.58.0
grpc-google-iam-v1==0.12.6
grpcio==1.51.3
grpcio-status==1.51.3
h11==0.14.0
httptools==0.5.0
idna==3.4
jmespath==1.0.1
libcst==0.4.9
mangum==0.17.0
mypy-extensions==1.0.0
proto-plus==1.22.2
protobuf==4.22.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydantic==1.10.6
PyMySQL==1.0.3
python-dateutil==2.8.2
python-dotenv==1.0.0
PyYAML==6.0
requests==2.28.2
rsa==4.9
s3transfer==0.6.0
six==1.16.0
sniffio==1.3.0
starlette==0.26.1
typing-extensions==4.5.0
typing-inspect==0.8.0
urllib3==1.26.15
uvicorn==0.21.0
uvloop==0.17.0
watchfiles==0.18.1
websockets==10.4
